Merge develop into main (#24) (#25)
* fix(config): fix esm build config

#6

* refactor(config): revert build config to commonjs

* feat(generation): add functions to generate package folder

#3

* feat(config): add basic config handling

A configuration file for package creation can now be red and the package creation location can be
modified

#11

* refactor(helpers): split helpers file

* feat(files): add sample files handling

Add all sample files into the new package

#4

* fix(files): add missing chalk in error

* fix(packagecreation): fix returned path

#18

* feat(files): handle multiple package types

If some package types are defined in the configuration the command will ask which type of package
should be created and will only copy the files in the package type folder. If there are no package
type defined, the command will copy all the files

* refactor(generatepackage): move the generate package code into helpers

In order to export/import the generate package code a new function as been created. The generation
package function can also take the relative path of a config file as an optional paramater to allow
the direct call of the function in the code

#21

* docs(readme): add readme file

#16